python/requirements.txt
#python 2.7
pandas

#imgaug requirements
six
numpy
scipy
Pillow
matplotlib
scikit-image
opencv-python
imageio Shapely
imgaug
